+++ src/object.ts
@@ -1,0 +1,28 @@
+/**
+ * @since 2.10.0
+ */
+import { getObjectSemigroup, Semigroup } from './Semigroup'
+
+// -------------------------------------------------------------------------------------
+// instances
+// -------------------------------------------------------------------------------------
+
+/**
+ * Return a semigroup which works like `Object.assign`.
+ *
+ * @example
+ * import { getAssignSemigroup } from 'fp-ts/object'
+ *
+ * interface Person {
+ *   readonly name: string
+ *   readonly age: number
+ * }
+ *
+ * const S = getAssignSemigroup<Person>()
+ * assert.deepStrictEqual(S.concat({ name: 'name', age: 23 }, { name: 'name', age: 24 }), { name: 'name', age: 24 })
+ *
+ * @category instances
+ * @since 2.10.0
+ */
+// tslint:disable-next-line: deprecation
+export const getAssignSemigroup: <A extends object = never>() => Semigroup<A> = getObjectSemigroup