+++ test/Semigroup.ts
@@ -32,25 +32,6 @@ describe('Semigroup', () => {
     U.deepStrictEqual(_.max(N.Ord).concat(1, 2), 2)
   })
 
-  it('assign', () => {
-    type T = {
-      readonly foo?: number
-      readonly bar: string
-    }
-    const foo: T = {
-      foo: 123,
-      bar: '456'
-    }
-    const bar: T = {
-      bar: '123'
-    }
-    const S = _.assign<T>()
-    const result = S.concat(foo, bar)
-    const expected = Object.assign({}, foo, bar)
-    U.deepStrictEqual(result.foo, expected.foo)
-    U.deepStrictEqual(result.bar, expected.bar)
-  })
-
   it('first', () => {
     U.deepStrictEqual(_.first<number>().concat(1, 2), 1)
   })